Updates to how panel plots are made
To maximize comprehensibility and ease of viewing, I now show each bad calibration as a panel plot of g, x, y, and z, with all good calibrations (earlier and later) in cyan and just a single bad calibration in black. The date and orientation of the bad calibration are given in the title and incorporated into the file names of the prints.