sensor reposition, easier comparison plots, better drift characterization
Many changes, most trivial.

The SCTA has been moved to a more stable position at Central Caldera and was found to be oriented with x~N and y~W. Lily-SCTA comparison code has been updated to reflect this.

Due to another interval of elevated noise on the Axial SCTA, I updated the calibration comparison code 'check_flips.m' to work with smaller pieces of data immediately around the time of the flip, resulting in faster operation and smaller file sizes.